# docs/08-api-spec.yaml
openapi: 3.0.3
info:
  title: SGRC — Sistema de Gestión y Reserva de Computadoras Educativas
  version: "1.0.0"
  description: |
    API mono-institucional para gestión de inventario y reservas de equipos
    informáticos. Monolito modular — un solo servicio, un solo dominio de datos.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://sgrc.tuinstitucion.edu.ar
    description: Producción
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT propio del sistema, emitido por `/api/auth/login` o `/api/auth/google`.

        **Todo 401 por sesión trae el header `X-Sesion-Motivo`** diciendo por
        qué no valió el token: `expirada` (venció por su `exp`), `invalida`
        (firma, algoritmo o formato), `revocada` (la cuenta ya no está
        habilitada) o `password-cambiada` (el token es anterior al último
        cambio de contraseña). Los cuatro son 401 y en los cuatro hay que
        descartar el token; el motivo sirve para decidir qué se le dice a la
        persona: un vencimiento no se anuncia, una sesión cerrada desde afuera
        sí. Va en `Access-Control-Expose-Headers`. Ver `09-seguridad-rbac.md` §4.

  # Paginación (internal/shared/paginacion). Solo la usan los listados que
  # crecen con el uso: reservas, notificaciones, usuarios, el buzón y el
  # registro de auditoría. Los demás —ciclos, cursos, materias, carros,
  # equipos— están acotados por el propio dominio (los carga un Admin a mano)
  # y se devuelven completos.
...
  schemas:
...
    Carro:
      type: object
      properties:
        dadoDeBaja: { type: boolean, description: "Baja lógica: el carro se retiró. No aparece en el listado." }
        fechaBaja: { type: string, format: date-time, nullable: true }
        id: { type: string, format: uuid }
        nombre: { type: string }
        descripcion: { type: string, nullable: true }
    Equipo:
      type: object
      properties:
        id: { type: string, format: uuid }
        etiqueta: { type: string, example: "PC 27", description: "Cómo se lo nombra en pantallas y correos: 'PC {identificador}' si está en un carro, el nombre si no. La resuelve el servidor para que la misma cosa no se vea distinta según dónde se la mire (RF-03.17)" }
        tipo: { type: string, maxLength: 50, example: "PROYECTOR", description: "Texto libre. 'PC' en las computadoras de un carro" }
        nombre: { type: string, maxLength: 100, nullable: true, example: "Proyector Epson", description: "Solo en equipos sueltos: es lo único que los distingue, y es único entre ellos sin distinguir mayúsculas. Ausente en las computadoras de un carro" }
        reservable: { type: boolean, description: "Si aparece en la lista de equipos libres al reservar (RF-03.16). Un proyector sí; un cargador no, se presta en el momento" }
        carroId: { type: string, format: uuid, nullable: true, description: "Ausente en un equipo suelto: un proyector no está en ningún carro" }
        identificador: { type: integer, nullable: true, description: "Único dentro del carro, no globalmente — puede repetirse en otro carro. Ausente en un equipo suelto", example: 27 }
        numeroSerie: { type: string, maxLength: 50, nullable: true, example: "5CD1234ABC", description: "Número de serie de fábrica, único en toda la institución. Alfanumérico: se guarda en mayúsculas y sin espacios al borde, así que la respuesta puede diferir de lo enviado. Ausente en un equipo suelto: un cargador puede no traer serie" }
        freezado: { type: boolean, description: "Informativo: indica si tiene Deep Freeze instalado. Sin efecto funcional sobre reservas." }
        cpu: { type: string, nullable: true }
        ram: { type: string, nullable: true }
        sistemaOperativo: { type: string, nullable: true }
        softwareInstalado: { type: string, nullable: true }
        estado: { type: string, enum: [DISPONIBLE, EN_MANTENIMIENTO, FUERA_DE_SERVICIO] }
        dadoDeBaja: { type: boolean, description: "Soft delete: si es true, no aparece en listados activos ni puede reservarse" }
        fechaBaja: { type: string, format: date-time, nullable: true }
        tieneCuentas: { type: boolean, description: "Si el equipo tiene anotada al menos una cuenta de usuario (RF-03.22). No se guarda: lo calcula la consulta. Sirve para no ofrecer 'Cómo entrar' donde no hay nada que ver — el tipo es texto libre, así que el sistema no puede deducir que un cargador no tiene con qué entrar, pero sí sabe que no tiene ninguna cuenta" }
...
paths:
...
  /api/carros:
    get:
      tags: [inventory]
      summary: Listar carros (cualquier usuario autenticado)
      description: |
        Visible para ADMIN y DOCENTE por igual — un docente necesita ver estado,
        freezado y softwareInstalado (ej. versión de AutoCAD) de cada PC para elegir
        cuáles reservar.

        Por defecto trae sólo los que están en circulación: esta misma lista
        llena el selector de "a qué carro va este equipo", y un carro retirado
        no es un destino válido.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: incluirRetirados
          required: false
          schema: { type: boolean }
          description: |
            Suma los carros dados de baja, con `dadoDeBaja: true`. Es lo único
            que hace alcanzable un carro retirado, y por lo tanto la única forma
            de llegar a reactivarlo. **Sólo para ADMIN**: para cualquier otro rol
            el parámetro se ignora.
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Carro' } }
    post:
      tags: [inventory]
      summary: Crear carro (Admin)
      description: |
        El nombre no puede repetirse, con la regla de unicidad de texto que vale
        en todo el sistema (RF-00.1): la comparación ignora tildes, mayúsculas y
        los espacios de más, así que en una institución que ya tiene «Carro
        EDUTEC», los nombres «CARRO EDUTEC», «carro edutec» y «Carro  EDUTEC»
        —con dos espacios— devuelven todos 409. Lo garantiza el índice único
        `ux_carro_nombre` sobre `clave_texto(nombre)`.

        El nombre se guarda canonizado: «  Carro   1  » queda como «Carro 1».
        La descripción NO se canoniza —no identifica nada y puede tener varios
        renglones a propósito—, sólo se le recortan los bordes.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre: { type: string, maxLength: 100, example: "Carro EDUTEC" }
                descripcion: { type: string, example: "Notebooks del laboratorio" }
      responses:
        "201": { description: Carro creado }
        "400": { description: "El nombre está vacío o pasa los 100 caracteres" }
        "409": { description: Ya existe un carro con ese nombre, sin distinguir tildes, mayúsculas ni espacios de más }
...
  /api/equipos:
    get:
      tags: [inventory]
      summary: Listar el inventario, entero o solo lo que no está en ningún carro (RF-03.15, todos los usuarios autenticados)
      description: |
        Sin filtro devuelve todo el inventario, que es lo que necesita una
        pantalla que ofrezca lo prestable sin recorrer carro por carro.

        Con `enCarro=false`, solo el proyector, los cargadores, las notebooks
        sueltas. Solo se reconoce el valor exacto `false`: cualquier otra cosa
        devuelve el inventario entero, porque de un filtro mal escrito es
        mejor ver de más que de menos — ver de menos se confunde con "no hay
        ninguno".

        Lo ve cualquier autenticado por el mismo motivo que los carros y sus
        equipos (RF-03.7): un docente necesita saber que existe un proyector
        antes de pedirlo.

        Los equipos sueltos comparten entidad —y esquema de respuesta— con las
        computadoras de carro, y eso es deliberado: "qué hay afuera del
        laboratorio" (RF-08) tiene que ser una sola lista.

        La condición va en la query y no en el path: un segmento literal como
        `/equipos/sueltos` obliga a registrarlo antes que `/{id}` o el
        parámetro se lo traga, y es una trampa que compila, arranca y falla en
        tiempo de ejecución.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: enCarro
          required: false
          schema: { type: string, enum: ["false"] }
          description: "`false` deja solo los equipos que no pertenecen a ningún carro."
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Equipo' }
    post:
      tags: [inventory]
      summary: Registrar un equipo suelto (RF-03.15, Admin)
      description: |
        A qué colección se hace POST decide dónde nace el equipo: acá nace
        suelto, en `/carros/{carroId}/equipos` nace adentro de ese carro.

        Son dos colecciones y no una con un campo opcional porque lo que se
        carga es distinto: acá no hay carro ni identificador, y lo que
        identifica al equipo es el nombre.

        El número de serie es **opcional** acá y obligatorio en una
        computadora de carro. No se divide en "notebooks sí, otros no" porque
        esa división es falsa: un proyector tiene serie —y es de lo que más se
        extravía— y un cargador no trae ninguna.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tipo, nombre]
              properties:
                tipo: { type: string, maxLength: 50, example: "PROYECTOR", description: "Texto libre: la lista de cosas que presta una escuela no es la de otra. El formulario sugiere los tipos ya cargados" }
                nombre: { type: string, maxLength: 100, example: "Proyector Epson", description: "Único entre los equipos sueltos, sin distinguir mayúsculas" }
                numeroSerie:
                  type: string
                  maxLength: 50
                  example: "ABC-123X"
                  description: |
                    Opcional, para cualquier tipo de equipo. Se guarda en
                    mayúsculas y sin espacios al borde, y es único entre TODOS
                    los equipos. Ausente o vacío se guarda como NULL, y una
                    columna `UNIQUE` admite tantos NULL como haga falta: veinte
                    cargadores sin serie conviven sin chocar.
                reservable: { type: boolean, default: false, description: "Marcar solo lo que alguien pueda querer planificar. Por defecto false: un cargador se presta en el momento y aparecería como ruido cada vez que un docente va a reservar" }
      responses:
        "201":
          description: Equipo registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Equipo' }
        "400": { description: Falta el tipo o el nombre, o exceden el largo }
        "409": { description: Ya hay otro equipo suelto con ese nombre }
...
  /api/incidencias:
    post:
      tags: [inventory]
      summary: Registrar incidencia sobre un equipo
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [equipoId, descripcion, gravedad]
              properties:
                equipoId: { type: string, format: uuid }
                descripcion: { type: string }
                categoria:
                  type: string
                  maxLength: 50
                  description: |
                    Opcional (RF-03.5): quien reporta no siempre sabe qué
                    falla. Texto libre; conviene tomarla de
                    `/api/categorias-de-falla` para no crear
                    variantes de una categoría que ya existe.
                  example: batería
                gravedad: { type: string, enum: [LEVE, MODERADA, GRAVE] }
      responses:
        "201": { description: Incidencia registrada }
        "400": { description: La categoría supera los 50 caracteres }
...
  /api/incidencias/equipos:
    get:
      tags: [reports]
      summary: Incidencias agregadas por equipo (RF-06.3, Admin)
      description: |
        Devuelve el CONTEO por equipo, no las incidencias sueltas. Incluye lo
        que no está en ningún carro (015): al proyector también se le rompe la
        lámpara.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: desde
          schema: { type: string, format: date }
        - in: query
          name: hasta
          schema: { type: string, format: date }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        equipoId: { type: string, format: uuid }
                        etiqueta: { type: string, example: "PC 3", description: "Cómo se nombra el equipo (RF-03.17). Es lo que se muestra: los dos de abajo van en 0 y \"\" si no está en ningún carro" }
                        identificador: { type: integer }
                        carroNombre: { type: string }
                        total: { type: integer }
                        abiertas: { type: integer }
                        enReparacion: { type: integer }
                        enviadasASoporte: { type: integer }
                        resueltas: { type: integer }
                        graves: { type: integer }

  # ══════════════════════════════
  # AVAILABILITY
  # ══════════════════════════════
...
  /api/incidencias/carros:
    get:
      tags: [reports]
      summary: Incidencias agregadas por carro (RF-06.3, Admin)
      description: |
        No depende del ciclo lectivo: Incidencia nunca se elimina (sobrevive
        al archivado, RF-02.4), así que se resuelve siempre en vivo.

        Este reporte agrupa POR CARRO, así que los equipos sueltos (015) no
        aparecen: no pertenecen a ninguno. Sus incidencias se ven en
        `/api/incidencias/equipos`.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: desde
          schema: { type: string, format: date }
        - in: query
          name: hasta
          schema: { type: string, format: date }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        carroId: { type: string, format: uuid }
                        carroNombre: { type: string }
                        total: { type: integer }
                        abiertas: { type: integer }
                        graves: { type: integer }
...